services layer added  and helper functions
